+++ History.md
@@ -3,6 +3,7 @@
  * [new] get buildlog preview by default instead of datestamp
  * [new] always show datestamp information below the file title
  * [new] add page commands which documents the tasks interface of Deb-o-Matic
+ * [new] better tooltip on action buttons (such as Download, View al file, Get all)
 
 # 0.4.2 (2014-06-15)
  * [fix] calculate pidfile for debomatic status check after merging configurations